requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.9.0
django-filter==21.1
djangorestframework==3.12.4
greenlet==1.1.2
gunicorn==20.1.0
numpy==1.21.2
pandas==1.3.3
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.25
sqlparse==0.4.2
